requirements.txt
asgiref==3.2.10
cairocffi==1.2.0
CairoSVG==2.5.0
cffi==1.14.4
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.1
django-quill-editor==0.1.16
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
html5lib==1.1
pdfkit==0.6.1
Pillow==8.0.1
pip-upgrade==0.0.6
pycparser==2.20
Pyphen==0.10.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
tinycss2==1.1.0
WeasyPrint==52.2
webencodings==0.5.1
